feat: rename to icrc candid
