+++ packages/ledger-icrc/src/index.canister.ts
@@ -2,9 +2,9 @@ import { Canister, createServices } from "@dfinity/utils";
 import type {
   GetTransactions,
   _SERVICE as IcrcIndexService,
-} from "../candid/icrc1_index";
-import { idlFactory as certifiedIdlFactory } from "../candid/icrc1_index.certified.idl";
-import { idlFactory } from "../candid/icrc1_index.idl";
+} from "../candid/icrc_index";
+import { idlFactory as certifiedIdlFactory } from "../candid/icrc_index.certified.idl";
+import { idlFactory } from "../candid/icrc_index.idl";
 import { toGetTransactionsArgs } from "./converters/index.converters";
 import { IndexError } from "./errors/index.errors";
 import type { IcrcLedgerCanisterOptions } from "./types/canister.options";
